Move all .h files to include & update paths
