+++ include/IndexEntry.h
