+++ include/vectorBook.h
@@ -46,6 +46,6 @@ private:
     void push(const vectorClass<T> &other);
 };
 
-#include "vectorBook.tpp"  // Include your implementation file
+#include "../src/vectorBook.tpp"  // Include your implementation file
 
 #endif // VECTORCLASS_H
