+++ src/vectorBook.tpp
@@ -1,6 +1,6 @@
 #include <iostream>
 #include <stdexcept>  // For std::out_of_range and std::runtime_error
-#include "vectorBook.h"
+#include "../include/vectorBook.h"
 
 using namespace std;
